Admin header helper added. Code for checking login code duplication added. Add to cart function fixed

+++ app/src/main/java/com/example/pos/data/dao/CartItemDao.kt
@@ -13,5 +13,5 @@ interface CartItemDao {
     @Delete
     suspend fun delete(cartItem: CartItem)
     @Query("SELECT * FROM cart WHERE oder_number = :orderNumber")
-    fun getCart(orderNumber: Int): LiveData<List<CartItem>>
+    fun getCart(orderNumber: Long): LiveData<List<CartItem>>
 }
